% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate_MAPE.R
\name{Calculate_MAPE}
\alias{Calculate_MAPE}
\title{Calculate MAPE}
\usage{
Calculate_MAPE(fit1, df_ts, orig, h)
}
\arguments{
\item{fit1}{The time series under consideration}

\item{df_ts}{The time series under consideration}

\item{orig}{This is the starting forecast origin. It should be less than the length of the underlying time series}

\item{h}{The forecast horizon. For a given h, it computes 1-step to h-step ahead forecasts}
}
\value{

}
\description{
This is a function help us calculate MAPE for time series in the last few rows after the RMSE and MAE Values
}
\examples{
Calculate_MAPE(fit1,df_ts,orig,h)
}
