% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_likelihood.R
\name{Log_likelihood}
\alias{Log_likelihood}
\title{Define a Log_likelihood Function}
\usage{
Log_likelihood(params, data)
}
\arguments{
\item{params}{This is a matrix with five params:p, mu1, sigma1, mu2, sigma2.}

\item{data}{This a observation data}
}
\value{

}
\description{
Define a log_likelihood Function with a matrix
}
\examples{
Log_likelihood(params, data)
}
