% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean_for_Colms.R
\name{Mean_for_Colms}
\alias{Mean_for_Colms}
\title{Return Mean for Each Columns and Drop NAs}
\usage{
Mean_for_Colms(x, removeNAs = TRUE)
}
\arguments{
\item{x}{This is the structured dataframe with integers or floats}

\item{removeNAs}{This equals TRUE means dropping NAs value}
}
\value{

}
\description{
Calculating mean for each columns and drop NAs at the same time
}
\examples{
Mean_for_Colms(x, removeNAs = devTRUE)
}
