% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Plot histogram}
\usage{
plot_hist(x, y, type = "1")
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a single plotting structure, function or any R object with a plot method can be provided}

\item{y}{the y coordinates of points in the plot, optional if x is an appropriate structure.}

\item{type}{input "1" means plot histogram (only accept 1)}
}
\value{

}
\description{
Help you plot a histogram directly
}
\examples{
plot_hist(x, y, type = "1")
}
