% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Todays_Date.R
\name{Todays_Date}
\alias{Todays_Date}
\title{Today's date with two types}
\usage{
Todays_Date(type = "")
}
\arguments{
\item{type}{This is the type of today's date you want to print.
And there are two types:
long - date with day of the week and month/day/year Eg.Thursday December 17 2020
short - simple date with month/day/year Eg.12/17/20}
}
\value{

}
\description{
Print today's date with two types
}
\examples{
Todays_Date("short")
}
